[FIX] Rename team not working properly (#21552)
